Move convertToMarketType helper to common file to remove duplication
